intercom-gpt-integration/requirements.txt
requests==2.26.0
python-dotenv==0.19.1
schedule==1.1.0  # For scheduling API polling
redis==4.1.0     # Optional: For caching and session storage
coverage==7.3.2  # For test coverage reporting
flask==1.1.4     # For webhook server (older version for better compatibility)
werkzeug==1.0.1  # Required by Flask 1.1.4
jinja2<3.0       # Required by Flask 1.1.4
markupsafe==1.1.1  # Older version with soft_unicode
itsdangerous<2.0 # Required by Flask 1.1.4
google-cloud-secret-manager==2.12.0  # For Secret Manager access
google-cloud-logging==3.1.2  # For Cloud Logging integration
gunicorn==20.1.0  # For production web server
